+++ packages/core/src/react/stores/plate/createPlateStore.ts
@@ -27,6 +27,7 @@ export const createPlateStore = <E extends PlateEditor = PlateEditor>({
   readOnly = null,
   renderElement = null,
   renderLeaf = null,
+  scrollRef = { current: null },
   versionDecorate = 1,
   versionEditor = 1,
   versionSelection = 1,
@@ -46,6 +47,7 @@ export const createPlateStore = <E extends PlateEditor = PlateEditor>({
       readOnly,
       renderElement,
       renderLeaf,
+      scrollRef,
       versionDecorate,
       versionEditor,
       versionSelection,
@@ -91,10 +93,6 @@ export const usePlateEditorStore = (
   // Try to fetch the store from a Plate provider
   const localStore = usePlateStore(id).store({ warnIfNoStore: false }) ?? null;
 
-  /**
-   * To preserve hook order, only use `localStore` if it was present on first
-   * render. This lets us call `usePlateControllerEditorStore` conditionally.
-   */
   const [localStoreExists] = React.useState(!!localStore);
 
   // If no store was found, try to fetch the store from a PlateController