+++ packages/heading/src/react/hooks/useTocSideBar.ts
@@ -3,9 +3,9 @@ import React from 'react';
 import { getNode } from '@udecode/plate-common';
 import {
   toDOMNode,
-  useEditorContainerRef,
   useEditorPlugin,
   useEditorSelector,
+  useScrollRef,
 } from '@udecode/plate-common/react';
 
 import type { Heading } from '../../lib/types';
@@ -23,7 +23,7 @@ export const useTocSideBarState = ({
 }: TocSideBarProps) => {
   const { editor } = useEditorPlugin(TocPlugin);
   const headingList = useEditorSelector(getHeadingList, []);
-  const containerRef = useEditorContainerRef();
+  const containerRef = useScrollRef();
 
   const tocRef = React.useRef<HTMLElement>(null);