+++ packages/markdown/src/lib/remark-slate/remarkDefaultElementRules.ts
@@ -11,8 +11,8 @@ export const remarkDefaultElementRules: RemarkElementRules = {
     transform: (node, options) => {
       const children = node.children?.length
         ? node.children.flatMap((paragraph) =>
-            remarkTransformElementChildren(paragraph, options)
-          )
+          remarkTransformElementChildren(paragraph, options)
+        )
         : [{ text: '' }];
 
       // Flatten nested blockquotes (e.g. >>>)
@@ -78,7 +78,7 @@ export const remarkDefaultElementRules: RemarkElementRules = {
           listItems: TElement[] = [],
           indent = 1
         ) => {
-          _node.children!.forEach((listItem) => {
+          _node.children?.forEach((listItem) => {
             const [paragraph, ...subLists] = listItem.children!;
 
             listItems.push({
@@ -114,12 +114,12 @@ export const remarkDefaultElementRules: RemarkElementRules = {
         };
 
         return parseListItems(node);
-      } else {
-        return {
-          children: remarkTransformElementChildren(node, options),
-          type: options.editor.getType({ key: node.ordered ? 'ol' : 'ul' }),
-        };
       }
+
+      return {
+        children: remarkTransformElementChildren(node, options),
+        type: options.editor.getType({ key: node.ordered ? 'ol' : 'ul' }),
+      };
     },
   },
   listItem: {
@@ -172,6 +172,36 @@ export const remarkDefaultElementRules: RemarkElementRules = {
       flushInlineNodes();
 
       return elements;
+    },
+  },
+  table: {
+    transform: (node, options) => {
+      const rows =
+        node.children?.map((row) => {
+          return {
+            children:
+              row.children?.map((cell) => {
+                return {
+                  type: options.editor.getType({ key: 'td' }),
+                  children: remarkTransformElementChildren(cell, options).map(child => {
+                    if (!child.type) {
+                      return {
+                        type: options.editor.getType({ key: 'p' }),
+                        children: [child]
+                      }
+                    }
+                    return child
+                  }),
+                };
+              }) || [],
+            type: options.editor.getType({ key: 'tr' }),
+          };
+        }) || [];
+
+      return {
+        children: rows,
+        type: options.editor.getType({ key: 'table' }),
+      };
     },
   },
   thematicBreak: {