+++ packages/basic-elements/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-basic-elements
 
+## 40.0.2
+
+## 40.0.1
+
 ## 40.0.0
 
 ## 39.0.0