+++ packages/plate-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @udecode/plate-utils
 
+## 40.0.3
+
+## 40.0.2
+
+## 40.0.1
+
 ## 40.0.0
 
 ### Minor Changes